+++ src/main/java/roomescape/reservationtime/service/ReservationTimeService.java
@@ -4,22 +4,35 @@ import java.time.LocalDate;
 import java.util.List;
 import java.util.stream.Collectors;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+import roomescape.exception.ErrorCode;
+import roomescape.exception.business.BusinessException;
 import roomescape.reservationtime.domain.ReservationTime;
+import roomescape.reservationtime.domain.ReservationTimeFactory;
 import roomescape.reservationtime.dto.TimeRequest;
 import roomescape.reservationtime.dto.TimeResponse;
 import roomescape.reservationtime.repository.ReservationTimeRepository;
 
 @Service
+@Transactional(readOnly = true)
 public class ReservationTimeService {
 
     private final ReservationTimeRepository timeRepository;
+    private final ReservationTimeFactory reservationTimeFactory;
 
-    public ReservationTimeService(ReservationTimeRepository timeRepository) {
+    public ReservationTimeService(ReservationTimeRepository timeRepository, ReservationTimeFactory reservationTimeFactory) {
         this.timeRepository = timeRepository;
+        this.reservationTimeFactory = reservationTimeFactory;
     }
 
+    public ReservationTime getById(Long id) {
+        return timeRepository.findById(id)
+                .orElseThrow(() -> new BusinessException(ErrorCode.TIME_NOT_FOUND));
+    }
+
+    @Transactional
     public TimeResponse createTime(TimeRequest request) {
-        ReservationTime time = ReservationTime.of(request.startAt(), request.finishAt());
+        ReservationTime time = reservationTimeFactory.create(request.startAt(), request.finishAt());
         ReservationTime saved = timeRepository.save(time);
         return TimeResponse.of(saved);
     }
@@ -36,7 +49,11 @@ public class ReservationTimeService {
                 .collect(Collectors.toList());
     }
 
+    @Transactional
     public void deleteById(Long id) {
+        if (timeRepository.existsReservationByTimeId(id)) {
+            throw new BusinessException(ErrorCode.TIME_HAS_RESERVATION);
+        }
         timeRepository.deleteById(id);
     }
 }
